created a separate class, which handles API requests and results
