+++ src/Controller/CocktailController.php
@@ -4,27 +4,14 @@ namespace App\Controller;
 
 use App\Form\CocktailType;
 use App\Entity\CocktailName;
-use http\QueryString;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpClient\HttpClient;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Contracts\HttpClient\HttpClientInterface;
-use function MongoDB\BSON\toJSON;
-use function PHPUnit\Framework\isNull;
 
 class CocktailController extends AbstractController
 {
-    private HttpClientInterface $httpClient;
-
-    private const URL = 'https://www.thecocktaildb.com/api/json/v1/1/search.php';
-
-    public function __construct(HttpClientInterface $httpClient)
-    {
-        $this->httpClient = $httpClient;
-    }
-
     #[Route('/', name: 'initForm')]
     public function initForm()
     {
@@ -64,30 +51,14 @@ class CocktailController extends AbstractController
 
         $cocktailName->setCocktailName($cocktailNameEntity->getCocktailName());
 
+        $client = HttpClient::create();
+        $parsedData = new CocktailApiController($client);
+
         //TODO API request action
 
         return $this->render('cocktail/submit.html.twig', [
             'cocktailForm' => $form->createView(),
-            'result' => $this->cocktailNameRequest($cocktailName)
+            'result' => $parsedData->cocktailNameRequest($cocktailName)
         ]);
-    }
-
-    private function cocktailNameRequest(CocktailName $cocktailName)
-    {
-        //TODO error logger
-
-        $response = $this->httpClient->request('GET', self::URL, [
-            'query' => [
-                's' => $cocktailName->getCocktailName()
-            ]
-        ]);
-
-        $parsedResponse = $response->toArray();
-
-        if(empty($parsedResponse['drinks']) || empty($parsedResponse['drinks'][0])){
-            return 'cocktail not found';
-        }
-
-        return $parsedResponse['drinks'];
     }
 }
